Format weekdays with i18n symbols too.

+++ generator/run.sh
@@ -2,8 +2,14 @@
 
 set -eu
 
+echo " [*] Install generator..."
 go install github.com/altipla-consulting/i18n-dateformatter/generator
-wget http://www.unicode.org/Public/cldr/27.0.1/core.zip -O /tmp/core.zip
 
+if [ ! -e /tmp/core.zip ]; then
+  echo " [*] Download CLDR data..."
+  wget http://www.unicode.org/Public/cldr/27.0.1/core.zip -O /tmp/core.zip
+fi
+
+echo " [*] Generate "
 generator -locales en,es,fr,ru
 gofmt -w symbols
